# GGRC-4926 Add access control list table optimizations