--- a/src/ggrc_workflows/models/hooks/workflow.py
+++ b/src/ggrc_workflows/models/hooks/workflow.py
@@ -3,8 +3,6 @@
 
 """SQLAlchemy hooks for Workflow model."""
 
-import logging
-
 import flask
 import sqlalchemy as sa
 
@@ -12,14 +10,8 @@ from ggrc import db
 from ggrc import login
 from ggrc import utils
 from ggrc.models import all_models
+from ggrc.models.hooks.acl import propagation
 from ggrc.access_control import role
-
-
-logger = logging.getLogger(__name__)
-
-
-# retry count for possible deadlock issues.
-PROPAGATION_RETRIES = 10
 
 
 RELATED_MODELS = (
@@ -137,60 +129,10 @@ def _get_child_ids(parent_ids, child_class):
 
 
 def _insert_select_acls(select_statement):
-  """Insert acl records from the select statement
-
-  Args:
-    select_statement: sql statement that contains the following columns
-      person_id,
-      ac_role_id,
-      object_id,
-      object_type,
-      created_at,
-      modified_by_id,
-      updated_at,
-      parent_id,
-      parent_id_nn,
-  """
-
+  """Run insert from select with ignore acl inserter."""
   acl_table = all_models.AccessControlList.__table__
-
   inserter = acl_table.insert().prefix_with("IGNORE")
-
-  last_error = None
-  for _ in range(PROPAGATION_RETRIES):
-    try:
-      last_error = None
-      db.session.execute(
-          inserter.from_select(
-              [
-                  acl_table.c.person_id,
-                  acl_table.c.ac_role_id,
-                  acl_table.c.object_id,
-                  acl_table.c.object_type,
-                  acl_table.c.created_at,
-                  acl_table.c.modified_by_id,
-                  acl_table.c.updated_at,
-                  acl_table.c.parent_id,
-                  acl_table.c.parent_id_nn,
-              ],
-              select_statement
-          )
-      )
-      db.session.plain_commit()
-      break
-    except sa.exc.OperationalError as error:
-      logger.exception(error)
-      last_error = error
-
-  if last_error:
-    logger.critical(
-        "Workflow ACL propagation failed with %d retries on statement: \n %s",
-        PROPAGATION_RETRIES,
-        select_statement,
-    )
-    # The following error if exists will only be sa.exc.OperationalError so the
-    # pylint warning is invalid.
-    raise last_error  # pylint: disable=raising-bad-type
+  propagation.insert_select_acls(inserter, select_statement)
 
 
 def _propagate_to_wf_children(new_wf_acls, child_class):