# Dependencies

This PR is `on hold` until the dependencies are merged:

- [x] https://github.com/google/ggrc-core/pull/7734
- [x] https://github.com/google/ggrc-core/pull/7773

I have added the https://github.com/google/ggrc-core/pull/7773 PR as a dependency because it has a conflict with this one. The conflict is already resolved here and this PR contains all changes.

# Issue description

Updating a single cycle task status takes more than 2 minutes, on bigger workflows (2k tasks and multiple cycles)

The biggest bottleneck in this is the query that pulls out all cycle tasks for a given cycle, which are then used for reindexing. The fix to not pull all tasks for the reindex job is done here: https://github.com/google/ggrc-core/issues?q=assignee%3Azidarsk8+is%3Aopen
But even so the query for multiple objects must never run that slow.

Example query that takes a few minutes (cut down version of the origina for readability):

```sql
SELECT                                                                          
  cycle_task_group_object_tasks.id AS cycle_task_group_object_tasks_id,            
  cycle_task_group_object_tasks.title AS cycle_task_group_object_tasks_title,   
  access_control_list_1.id AS access_control_list_1_id                          
FROM cycle_task_group_object_tasks                                              
LEFT OUTER JOIN access_control_list AS access_control_list_1 ON                 
  access_control_list_1.object_id = cycle_task_group_object_tasks.id AND        
  access_control_list_1.object_type = "CycleTaskGroupObjectTask" AND            
  access_control_list_1.parent_id IS NULL                                          
WHERE                                                                           
  cycle_task_group_object_tasks.cycle_id = 6662                                 
;                                                                               
```
results of the query above
```
...
2000 rows in set (1 min 9.76 sec)
```

Now the tinyest change in the query produces same results but in better time:
```sql
SELECT                                                                          
  cycle_task_group_object_tasks.id AS cycle_task_group_object_tasks_id,            
  cycle_task_group_object_tasks.title AS cycle_task_group_object_tasks_title,   
  access_control_list_1.id AS access_control_list_1_id                          
FROM cycle_task_group_object_tasks                                              
LEFT OUTER JOIN access_control_list AS access_control_list_1 ON                 
  access_control_list_1.object_id = cycle_task_group_object_tasks.id AND        
  access_control_list_1.object_type = "CycleTaskGroupObjectTask"                
WHERE                                                                           
  access_control_list_1.parent_id IS NULL AND                                   
  cycle_task_group_object_tasks.cycle_id = 6662   

```
results of the query above
```
...
2000 rows in set (0.06 sec)
```

Issue turns out to be invalid usage of index, even when a more propper one is created:
```sql
mysql> show create table access_control_list\G
*************************** 1. row ***************************
       Table: access_control_list
Create Table: CREATE TABLE `access_control_list` (
  `id` int(11) NOT NULL AUTO_INCREMENT,
  `person_id` int(11) NOT NULL,
  `ac_role_id` int(11) NOT NULL,
  `object_id` int(11) NOT NULL,
  `object_type` varchar(250) NOT NULL,
  `created_at` datetime NOT NULL,
  `modified_by_id` int(11) DEFAULT NULL,
  `updated_at` datetime NOT NULL,
  `context_id` int(11) DEFAULT NULL,
  `parent_id` int(11) DEFAULT NULL,
  PRIMARY KEY (`id`),
  UNIQUE KEY `person_id` (`person_id`,`ac_role_id`,`object_id`,`object_type`,`parent_id`),
  KEY `ac_role_id` (`ac_role_id`),
  KEY `fk_access_control_list_contexts` (`context_id`),
  KEY `ix_access_control_list_updated_at` (`updated_at`),
  KEY `idx_object_type_object_idx` (`object_type`,`object_id`),
  KEY `fk_access_control_list_parent_id` (`parent_id`),
  KEY `acl_join_index` (`object_type`,`object_id`,`parent_id`),
  CONSTRAINT `access_control_list_ibfk_1` FOREIGN KEY (`ac_role_id`) REFERENCES `access_control_roles` (`id`),
  CONSTRAINT `access_control_list_ibfk_2` FOREIGN KEY (`context_id`) REFERENCES `contexts` (`id`),
  CONSTRAINT `access_control_list_ibfk_3` FOREIGN KEY (`person_id`) REFERENCES `people` (`id`),
  CONSTRAINT `fk_access_control_list_parent_id` FOREIGN KEY (`parent_id`) REFERENCES `access_control_list` (`id`) ON DELETE CASCADE
) ENGINE=InnoDB AUTO_INCREMENT=5450921 DEFAULT CHARSET=utf8
1 row in set (0.00 sec)
```

The explain of the slow query:
```sql
+----+-------------+-------------------------------+------+----------------------------------------------------------------------------+-------------------------------------+---------+-------+------+-------------+
| id | select_type | table                         | type | possible_keys                                                              | key                                 | key_len | ref   | rows | Extra       |
+----+-------------+-------------------------------+------+----------------------------------------------------------------------------+-------------------------------------+---------+-------+------+-------------+
|  1 | SIMPLE      | cycle_task_group_object_tasks | ref  | cycle_task_group_object_tasks_cycle                                        | cycle_task_group_object_tasks_cycle | 4       | const |  999 | NULL        |
|  1 | SIMPLE      | access_control_list_1         | ref  | idx_object_type_object_idx,fk_access_control_list_parent_id,acl_join_index | fk_access_control_list_parent_id    | 5       | const |    1 | Using where |
+----+-------------+-------------------------------+------+----------------------------------------------------------------------------+-------------------------------------+---------+-------+------+-------------+
```
 
This shows that `parent_id` index is used instead of `acl_join_index`.

If we force the usage of the correct index, we get the expected performance:
```sql
mysql> SELECT                                                                                                                                                                                                                                         ->   cycle_task_group_object_tasks.id AS cycle_task_group_object_tasks_id,         
    ->   cycle_task_group_object_tasks.title AS cycle_task_group_object_tasks_title,   
    ->   access_control_list_1.id AS access_control_list_1_id                          
    -> FROM cycle_task_group_object_tasks                                              
    -> LEFT OUTER JOIN access_control_list AS access_control_list_1                    
    -> FORCE INDEX FOR JOIN (`acl_join_index`)                                         
    -> ON                                                                              
    ->   access_control_list_1.object_id = cycle_task_group_object_tasks.id AND        
    ->   access_control_list_1.object_type = "CycleTaskGroupObjectTask" AND            
    ->   access_control_list_1.parent_id IS NULL                                       
    -> WHERE                                                                           
    ->   cycle_task_group_object_tasks.cycle_id = 6662                                 
    -> ;
+----------------------------------+-------------------------------------+--------------------------+
| cycle_task_group_object_tasks_id | cycle_task_group_object_tasks_title | access_control_list_1_id |
+----------------------------------+-------------------------------------+--------------------------+
|                            36236 | big-wf-2-tgt-1big-wf-2-tg-1         |                  1340186 |
...
+----------------------------------+-------------------------------------+--------------------------+
2000 rows in set (0.01 sec)

```


# Steps to test the changes

Modify a cycle task in a cycle with a lot of tasks and many people on the workflow.
Verify that the PUT request takes less than 5 seconds.
Verify that correct CycleTasks were pulled from the database by checking if CycleTask reindex works.

# Solution description

Because we can't (or I don't know how to) force the correct index on sqlalchemy relationship join, we are forced to remove the invalid index. But we still need the FK index for delete cascade.

The solution is then to just duplicate the FK column, that is then used in the query and that is properly indexed. 
The added benefit of this column is that by making it non null, wen can also prevent duplicate acl entries.


# Sanity checklist

- [x] I have clicked through the app to make sure my changes work and not break the app.
- [ ] I have applied the correct milestone and labels.
- [x] My changes fix the issue described in the description (and do nothing else). 🤞
- [x] My changes are covered by tests. - regressions are covered by previous tests, but there are no performance tests here.
- [ ] My changes follow our [performance guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/performance.rst). - The reindex code that pulls too much data was not addressed in this PR.
- [x] My changes follow our [js](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/javascript.rst) and/or [python](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/python.rst) guidelines.
- [x] My commits follow our [commit guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/git/how_to_write_a_commit_message.rst).

<!-- If your PR includes a migration include the additional checklist items
# Migration checklist
- [ ] Migration passes all checks from our [PR review guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/git/reviewing_pull_requests.rst#reviewing-a-pr-containing-database-migration-scripts)
-->

# PR Review checklist

- [x] The changes fix the issue and don't cause any apparent regressions.
- [x] Labels and milestone are correctly set.
- [x] The solution description matches the changes in the code.
- [x] There is no apparent way to improve the performance & design of the new code.
- [x] The pull request is opened against the correct base branch.
- [x] Upon merging, the Jira ticket's fixversion is correctly set and the ticket is moved to "QA - In Progress".

<!-- If your code is not finished yet can include a TODO check list
# TODO

- [ ] First item on the TODO
-->