@ahetmanski 2018-05-11
@zidarsk8 Codestyle failed here. Code duplicates were found.