@zidarsk8 2018-05-14
@ahetmanski Regarding the warning, it is an invalid warining that sqlachemy shows when running
`delete table if exists`, if you downgrade and upgrade the warning is gone. but it just shows that we are refferencing a table that does not yet exist, which is fine because we have the check for it in the delete clause.