@ahetmanski 2018-05-15
@zidarsk8 I am talking about next warning, and it occurs because of comment above. You [copy](https://github.com/google/ggrc-core/pull/7732/commits/9871f5dbc6be98b78e6deb749c5af10e30e6e556#diff-feec825352d7187314edb3309e1f5fb2R79) _null_ values from _parent_id_ to _parent_id_nn_, but last one is not nullable.

> Warning: Column 'parent_id_nn' cannot be null